#splitted data into stores
