+++ src/lib/stores/skills.ts
@@ -1,0 +1,20 @@
+import { readable } from "svelte/store";
+
+export const skills = readable<string[]>([
+  "Vue.js",
+  "Nuxt.js",
+  "Pinia",
+  "Svelte",
+  "TailwindCSS",
+  "Bootstrap",
+  "Shadcn",
+  "JavaScript",
+  "TypeScript",
+  "HTML",
+  "CSS",
+  "SCSS",
+  "Git",
+  "GitHub",
+  "Firebase",
+  "Vercel",
+]);
