+++ src/routes/+page.svelte
@@ -5,106 +5,19 @@
   import CustomInput from "$lib/components/CustomInput.svelte";
   import CustomLabel from "$lib/components/CustomLabel.svelte";
   import { fade } from "svelte/transition";
-  import { toast } from "$lib/stores/store";
+  import { toast } from "$lib/stores/toast";
+  import { skills } from "$lib/stores/skills";
+  import { experiences } from "$lib/stores/experiences";
+  import { projects } from "$lib/stores/projects";
   import Icon from "@iconify/svelte";
   import emailjs from "@emailjs/browser";
-  import type Experience from "$lib/types/Experience";
-  import type Project from "$lib/types/Project";
   import ogImage from "../../static/ogImage.png";
 
   const title: string = "Aderibigbe Michael O.";
   const description: string =
     "Innovative Frontend Developer with expertise in Vue.js. Enhancing user experiences and optimizing web performance. Skilled in developing responsive designs and resolving complex code issues. Implemented cutting-edge solutions, demonstrating technical proficiency and problem-solving abilities.";
 
-  const skillBadges: string[] = [
-    "Vue.js",
-    "Nuxt.js",
-    "Pinia",
-    "Svelte",
-    "TailwindCSS",
-    "Bootstrap",
-    "Shadcn",
-    "JavaScript",
-    "TypeScript",
-    "HTML",
-    "CSS",
-    "SCSS",
-    "Git",
-    "GitHub",
-    "Firebase",
-    "Vercel",
-  ];
-
-  let experiences: Experience[] = [
-    {
-      date: "May 2024 - Present",
-      title: "Frontend Developer, Stranerd",
-      description:
-        "Stranerd is an edu-tech company focused on digitalizing all aspects of education. At Stranerd, I build and maintain web components and layouts, optimize web applications, enhance landing pages, and integrate RESTful APIs to ensure real-time data flow.",
-      badges: ["Vuejs", "TypeScript", "TailwindCSS", "Figma"],
-      path: "https://stranerd.com",
-    },
-    {
-      date: "June 2023 - Present",
-      title: "Frontend Developer, AG GLOWSENSE",
-      description:
-        "AG-GLOWSENSE is a beauty, cosmetics, and skincare company specializing in delivering high-quality products. I developed and I am maintaining the full-stack e-commerce platform, focusing on optimizing frontend performance and responsiveness. I also implemented secure user authentication and data storage, while integrating the Gemini AI Chat Assistant to improve customer support.",
-      badges: [
-        "Vuejs",
-        "JavaScript",
-        "TailwindCSS",
-        "Pinia",
-        "Firebase",
-        "Vercel",
-        "Gemini AI",
-      ],
-      path: "https://ag-glowsense.vercel.app/",
-    },
-  ];
-  let projects: Project[] = [
-    {
-      imagePath: "/devlinks-img.png",
-      title: "Devlinks",
-      description:
-        "A profile-building app for developers to create and preview custom profiles. Shareable links enable easy sharing with others.",
-      skillSets: ["Nuxt 3", "Typescript", "Firebase", "Pinia", "TailwindCSS"],
-      urlPath: "https://devlinks-31.vercel.app/",
-    },
-    {
-      imagePath: "/kanban.png",
-      title: "Kanban Task Manager",
-      description:
-        "A task manager app for organizing tasks into boards with customizable columns where tasks can have description, subtasks.",
-      skillSets: ["Vue 3", "Typescript", "Firebase", "Pinia", "TailwindCSS"],
-      urlPath: "https://kanban-task-manager-31.vercel.app/",
-    },
-    {
-      imagePath: "/dashboard-img.png",
-      title: "Dashboard Clone",
-      description:
-        "An e-commerce dashboard clone displaying data analytics and product details using charts. Includes a theme switcher for an enhanced user experience.",
-      skillSets: ["Nuxt 3", "Shadcn", "TailwindCSS", "Fake Api", "ApexCharts"],
-      urlPath: "https://dashboard-clone-delta.vercel.app/",
-    },
-    {
-      imagePath: "/cinemax-img.png",
-      title: "Cinemax",
-      description:
-        "A movie app where users can discover trending movies and series, watch trailers, and search for specific titles to get detailed information.",
-      skillSets: ["Vue 3", "SCSS", "TailwindCSS", "Pinia", "TMDB API"],
-      urlPath: "https://cinemax-31.vercel.app/",
-    },
-    {
-      imagePath: "/bookmark-img.png",
-      title: "Bookmark Landing Page",
-      description:
-        "A minimalist landing page for a bookmarking app, designed for simplicity and ease of use.",
-      skillSets: ["Svelte", "TailwindCSS"],
-      urlPath: "https://bookmark-landing-blush.vercel.app/",
-    },
-  ];
-
-  //contact & sendMail
+    //contact & sendMail
   let emailInput: string = "";
   let fullNameInput: string = "";
   let messageInput: string = "";
@@ -212,7 +125,7 @@
       <div
         class="grid grid-cols-3 md:grid-cols-6 lg:grid-cols-3 xl:grid-cols-5 gap-3"
       >
-        {#each skillBadges as badge}
+        {#each $skills as badge}
           <Badge content={badge} />
         {/each}
       </div>
@@ -225,7 +138,7 @@
     <h2 class="uppercase font-bold">My Experiences</h2>
   </div>
   <div class="section-content">
-    {#each experiences as experience}
+    {#each $experiences as experience}
       <a href={experience.path} target="_blank">
         <Card content={experience} />
       </a>
@@ -251,7 +164,7 @@
     <h2 class="uppercase font-bold">My Projects</h2>
   </div>
   <div class="section-content">
-    {#each projects as project}
+    {#each $projects as project}
       <a href={project.urlPath} target="_blank">
         <ProjectCard content={project} />
       </a>
